Move verify vc
